+++ src/utils/id-util.js
@@ -1,0 +1,21 @@
+/**
+ * 生成一个随机id
+ *
+ * @returns {string}
+ */
+function randomId() {
+    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+        const r = Math.random() * 16 | 0,
+            v = c === 'x' ? r : (r & 0x3 | 0x8);
+        return v.toString(16);
+    });
+}
+
+module.exports = {
+    randomId
+}
+
+
+
+
+
