+++ userscript-headers.js
@@ -6,5 +6,8 @@
 // @document     ${document}
 // @author       ${author}
 // @match        *://*/*
-// @run-at       document-start
+// @require https://cdn.bootcss.com/jquery/1.12.4/jquery.min.js
+// @require      https://cdn.staticfile.org/jszip/3.5.0/jszip.min.js
+// @run-at document-idle
+// @grant GM_xmlhttpRequest
 // ==/UserScript==
